% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klippy.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_clipboard}
\alias{html_dependency_klippy}
\alias{html_dependency_octicon_clippy}
\title{Provide HTML dependencies}
\usage{
html_dependency_clipboard()

html_dependency_klippy()

html_dependency_octicon_clippy()
}
\value{
An object that can be included in a list of dependencies passed to
  \code{\link{attachDependencies}}.
}
\description{
These functions provide HTML dependencies for `clipboard.js`, `klippy` and
`octicon-clippy` for re-use.
}
