% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_dependencies.R
\name{html_dependencies}
\alias{html_dependencies}
\alias{html_dependency_clipboard}
\alias{html_dependency_primer_tooltips}
\alias{html_dependency_klippy}
\title{Provide HTML dependencies}
\usage{
html_dependency_clipboard()

html_dependency_primer_tooltips()

html_dependency_klippy()
}
\value{
An object that can be included in a list of dependencies passed to
\code{\link[htmltools:attachDependencies]{htmltools::attachDependencies()}}.
}
\description{
These functions provide HTML dependencies for \code{clipboard.js},
\code{klippy} and \verb{Primer Tooltips CSS} for re-use.
}
\seealso{
Other HTML dependencies functions: 
\code{\link{klippy_dependencies}()}
}
\concept{HTML dependencies functions}
