% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klippy.R
\name{klippy}
\alias{klippy}
\title{Insert copy to clipboard buttons in HTML documents}
\usage{
klippy(lang = "r markdown", all_precode = FALSE)
}
\arguments{
\item{lang}{A character string or a vector of character strings with
language names. If a character string contains multiple languages
names, these names have to be separated by boundaries (e.g., spaces).
Void string can be passed to \code{lang} argument.}

\item{all_precode}{A logical scalar. If \code{TRUE}, a "klippy" is
added to all \code{HTML <pre>} elements having an \code{HTML <code>}
element as a child.}
}
\value{
An \code{HTML \link[htmltools]{tag} object}.
}
\description{
\code{klippy} insert copy to clipboard buttons (or "klippies") in \code{R}
\code{Markdown} \code{HTML} documents. In the rendered document, "klippies"
are inserted in the upper right corner of the code chunks. \code{klippy()}
function is suited for a call in a \code{knitr} code chunk.
}
\details{
\code{klippy()} function appends \code{JavaScript} functions and \code{CSS} in
the rendered document that:
\enumerate{
\item Add \code{klippy} to the class attribute of selected \code{<pre>}
elements.
\item Add a \code{<button>} element (a "klippy") as a child for all
\code{<pre>} elements with a \code{klippy} class attribute.
\item Instantiate \code{clipboard.js} event listeners and attach them to
\code{klippies}.}
\code{klippy} class can also be appended to a \code{<pre>} element using
\href{https://yihui.name/knitr/options/}{\code{knitr class.source} chunk
option}. "Klippy" buttons are not rendered if the browser does not support
\code{clipboard.js} library (see \href{https://clipboardjs.com/}{here} for
details).
}
\section{HTML Dependencies}{

\code{klippy} package includes several third parties \code{JavaScript}
libraries that are declared as \code{HTML} dependencies:
\itemize{
  \item \code{jQuery-1.11.3}, (c) jQuery Foundation, Inc.
  \href{https://jquery.org/license/}{MIT license}.
  \href{https://jquery.com/}{Website}.
  \item \code{Bootstrap-3.3.5}, (c) Twitter, Inc.
  \href{https://github.com/twbs/bootstrap/blob/v3.3.5/LICENSE}{MIT license}.
  \href{http://getbootstrap.com}{Website}.
  \item \code{clipboard.js-1.7.1}, (c) Zeno Rocha
  \email{hi@zenorocha.com}.
  \href{http://zenorocha.mit-license.org/}{MIT license}.
  \href{https://clipboardjs.com/}{Website}.}
}

\examples{
tf <- tempfile(fileext = c(".Rmd", ".html"))
writeLines(
  c("```{r, echo=TRUE, include=TRUE}",
    "klippy::klippy()",
    "```"
  ),
  tf[1]
)
rmarkdown::render(tf[1], "html_document", tf[2])

\dontrun{
browseURL(paste0("file://", tf[2]))}

}
