% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klippy.R
\name{klippy_dependencies}
\alias{klippy_dependencies}
\title{List klippy dependencies}
\usage{
klippy_dependencies()
}
\value{
A list of dependencies that can be passed to
\code{\link[htmltools]{attachDependencies}} or
\code{\link[rmarkdown]{html_document_base}}.
}
\description{
This function is used to get the list of \code{klippy} dependencies.
}
\seealso{
Other HTML dependencies functions: \code{\link{html-dependencies}}
}
