% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klippy.R
\name{html-dependencies}
\alias{html-dependencies}
\alias{html_dependency_clipboard}
\alias{html_dependency_klippy}
\alias{html_dependency_octicon_clippy}
\title{Provide HTML dependencies}
\usage{
html_dependency_clipboard()

html_dependency_klippy()

html_dependency_octicon_clippy()
}
\value{
An object that can be included in a list of dependencies passed to
\code{\link[htmltools]{attachDependencies}}.
}
\description{
These functions provide HTML dependencies for \code{clipboard.js},
\code{klippy} and \code{octicon-clippy} for re-use.
}
\seealso{
Other HTML dependencies functions: \code{\link{klippy_dependencies}}
}
